DjangoShop/requirements.txt
django >=1.9, <2
dj-pagination
djangorestframework
markdown
django-filter